Refactoring n stuff
