+++ src/main/java/com/dfsek/terra/config/loaders/palette/PaletteHolderLoader.java
@@ -1,10 +1,11 @@
-package com.dfsek.terra.config.loaders;
+package com.dfsek.terra.config.loaders.palette;
 
 import com.dfsek.tectonic.exception.LoadException;
 import com.dfsek.tectonic.loading.ConfigLoader;
 import com.dfsek.tectonic.loading.TypeLoader;
 import com.dfsek.terra.biome.palette.PaletteHolder;
 import com.dfsek.terra.biome.palette.PaletteHolderBuilder;
+import com.dfsek.terra.config.loaders.Types;
 import org.bukkit.block.data.BlockData;
 import org.polydev.gaea.world.palette.Palette;
